+++ lib/msal-browser/test/custom_auth/integration_tests/ResetPassword.spec.ts
@@ -14,33 +14,7 @@ import { CustomAuthStandardController } from "../../../src/custom_auth/controlle
 import { ResetPasswordCodeRequiredState } from "../../../src/custom_auth/reset_password/auth_flow/state/ResetPasswordCodeRequiredState.js";
 import { ResetPasswordPasswordRequiredState } from "../../../src/custom_auth/reset_password/auth_flow/state/ResetPasswordPasswordRequiredState.js";
 import { ResetPasswordCompletedState } from "../../../src/custom_auth/reset_password/auth_flow/state/ResetPasswordCompletedState.js";
-
-jest.mock("@azure/msal-common/browser", () => {
-    const actualModule = jest.requireActual("@azure/msal-common/browser");
-    return {
-        ...actualModule,
-        ResponseHandler: jest.fn().mockImplementation(() => ({
-            handleServerTokenResponse: jest.fn().mockResolvedValue({
-                uniqueId: "test-unique-id",
-                tenantId: "test-tenant-id",
-                scopes: ["test-scope"],
-                account: {
-                    homeAccountId: "test-home-account-id",
-                    environment: "test-environment",
-                    tenantId: "test-tenant-id",
-                    username: "test-username",
-                    idToken: "test-id-token",
-                },
-                idToken: "test-id-token",
-                idTokenClaims: {},
-                accessToken: "test-access-token",
-                refreshToken: "test-refresh-token",
-                expiresOn: new Date(),
-                extExpiresOn: new Date(),
-            }),
-        })),
-    };
-});
+import { TestServerTokenResponse } from "../test_resources/TestConstants.js";
 
 describe("Reset password", () => {
     let app: CustomAuthPublicClientApplication;
@@ -152,16 +126,7 @@ describe("Reset password", () => {
             .mockResolvedValueOnce({
                 status: 200,
                 json: async () => {
-                    return {
-                        correlation_id: correlationId,
-                        token_type: "Bearer",
-                        scopes: "test-scope",
-                        expires_in: 3600,
-                        id_token: "test-id-token",
-                        access_token: "test-access-token",
-                        refresh_token: "test-refresh-token",
-                        client_info: "test-client-info",
-                    };
+                    return TestServerTokenResponse;
                 },
                 headers: new Headers({ "content-type": "application/json" }),
                 ok: true,
@@ -206,8 +171,11 @@ describe("Reset password", () => {
         expect(signInResult.data).toBeDefined();
         expect(signInResult.data).toBeInstanceOf(CustomAuthAccountData);
         expect(signInResult.data?.getAccount()?.idToken).toStrictEqual(
-            "test-id-token"
+            TestServerTokenResponse.id_token
         );
+
+        // Sign out the user for clean up the state for the other tests.
+        signInResult.data?.signOut();
     });
 
     it("should reset password failed if the redirect challenge returned", async () => {