+++ lib/msal-react/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @azure/msal-react
 
-<!-- This log was last generated on Wed, 18 Jun 2025 15:21:55 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 01 Jul 2025 14:23:25 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 3.0.14
+
+Tue, 01 Jul 2025 14:23:25 GMT
+
+### Patches
+
+- Bump @azure/msal-browser to v4.14.0
+- Bump eslint-config-msal to v0.0.0
+- Bump msal-test-utils to v0.0.1
 
 ## 3.0.13